Custom Themes
Here we have created a custom theme with "createMuiTheme" and utilized it by wrapping the rest of our App component's components within the "ThemeProvider" component with its "theme" prop set to the theme we created.

We also remove and comment out the makeStyles utilization in the prior commit.

See diff and related comments.