TextField component - App Implementation (3) - Validation error
Here we utilize the TextField components error prop to highlight a field in red that the user has not enter input into on form submit.

The error prop takes a boolean and we're now tracking a value for each text field on state.  During our handleSubmit method, I'm updating these values to true/false if the text field lacks/has a value and returning early if either is missing.