Radio Buttons - App Implementation (1) - Simple Radio Group
Here we add a simple group of radio buttons and track the state of the selected value with the useState hook.

This simple radio button group involves three Material UI components (RadioGroup, FormControlLabel and Radio).

(We will add the other two described in README.md and on Material UI radio docs page with next commit.)