Radio Buttons - App Implementation (2) - Add Group Label
Here we add the FormControl and FormLabel components.  The FormLabel allows us to put a label on the entire group.  And the FormControl component is like a wrapper that goes around FormLabel and RadioGroup as its direct children.

I also updated the primary color here to avoid the label from disappearing when the radio group is in focus.  For some reason, Material UI has a default behavior where this Label will change colors when a control within it is changed (pretty ugly imo.)

We're now using all 5 components like in the Material UI radio buttons documentation here: https://material-ui.com/components/radio-buttons/#radio