Grid System - Actual Implementation on Notes
We get rid of the practice content and apply the Material UI Grid system to our actual Notes. We also nest the whole thing in a Container component for a little margin.

The styling still doesn't look great overall but the responsive distribution of elements for different screen sizes responsively is now in place.