Card Component - Prep Work
Setting up a new NoteCard component to use in our existing Notes component, where we'll pass the note object as a prop.  For now, we just use that in NoteCard to display the note.title.

Next, we'll utilize Material UI's Card component in NoteCard to flesh it out and style it up nice.